Removing unnecessary code and added some quality of life changes.

+++ scr/WindowFunctions/WindowFunctions.cpp
@@ -7,7 +7,10 @@ void SpriteRenderer::Resize(GLFWwindow* window, int width, int height)
 	glViewport(0, 0, width, height);
 	*Global::winHeight = height;
 	*Global::winWidth = width;
-	*Global::projection = glm::ortho<float>(-*Global::winWidth / 2, *Global::winWidth / 2, -*Global::winHeight / 2, *Global::winHeight / 2, -1.0, 1.0);
+	*Global::projection = glm::ortho<float>((float)( - *Global::winWidth / 2),
+											(float)(*Global::winWidth / 2), 
+											(float)( - *Global::winHeight / 2), 
+											(float)(*Global::winHeight / 2), -1.0, 1.0);
 }
 
 void SpriteRenderer::key_callback(GLFWwindow* window, int key, int scancode, int action, int mods)
